==> analysis/03_01_Posthoc_analyses_Gene_Enrichment.Rmd <==
---
title: "The Methylome in Female Adolescent Conduct Disorder: Neural Pathomechanisms and Environmental Risk Factors"
subtitle: 'Post-hoc Analyses'
author: "AG Chiocchetti"
date: "29 Dezember 2020"
output: html_document
code_folding: hide
editor_options: 
  markdown: 
    wrap: 72
---

```{r}
Home = getwd()
```

```{r setup, echo=F, include=FALSE}
library(RColorBrewer)
library(corrplot)
library(gplots)
library(ggplot2)
library(WGCNA)
library(viridis)
library(lavaan)
library(semPlot)
library(tidyverse)
library(DESeq2)
library(GenomicRanges)
library(knitr)
library(glmpca)
library(kableExtra)
library(BiocParallel)
library(gprofiler2)
library(xlsx)
library(org.Hs.eg.db)
library(visNetwork)
library(webshot); #webshot::install_phantomjs() #in case phantomjs was not installed 
library(RRHO)
library(readxl)



thresholdp = 10e-3
thresholdLFC = 0.5


source(paste0(Home, "/code/custom_functions.R"))

options(stringsAsFactors = F)
opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dev = c("png"))


Dark8 = brewer.pal(8, "Dark2")
Set3 = brewer.pal(12, "Set3")
Dark8_50 = paste0(brewer.pal(8, "Dark2"), "7D")

jetcolors = colorRampPalette(c("darkblue", "skyblue", "green", "yellow", "orange", "red", "darkred"))
OBcolors = colorRampPalette(c("darkblue", "skyblue", "white",  "orange", "darkorange3"))
load(paste0(Home,"/output/dds_filt_analyzed.RData"))
load(paste0(Home,"/output/resdmr.RData"))
load(paste0(Home,"/output/resultsdmr_table.RData"))
load(paste0(Home,"/output/envFact.RData"))
load(paste0(Home,"/output/clinFact.RData"))

results_Deseq = results(dds_filt)
restab=results_Deseq
selEpitpm = counts(dds_filt, normalized=T)
log2_cpm = log2(counts(dds_filt, normalized=T)+1)
selEpiMeta = mcols(dds_filt)


```

# Sensitivity Analyses

## Sensitivity EWAS

### including SES

```{r effect_of_SES}
collector=data.frame(originalP=results_Deseq$pvalue,
                     originall2FC=results_Deseq$log2FoldChange)

rownames(collector)=paste0("Epi", 1:nrow(collector))

parm="EduPar"

workingcopy = dds_filt
workingcopy=workingcopy[,as.vector(!is.na(colData(dds_filt)[parm]))]
modelpar=as.character(design(dds_filt))[2]
tmpmod=gsub("0", paste0("~ 0 +",parm), modelpar)
tmpmod=gsub("int_dis \\+", "", tmpmod)

modelpar=as.formula(tmpmod)
design(workingcopy) = modelpar

workingcopy = DESeq(workingcopy)
parmres=results(workingcopy)

collector[,paste0(parm,"P")] = parmres$pvalue
collector[,paste0(parm,"l2FC")] = parmres$log2FoldChange

idx=collector$originalP<=thresholdp
idx=collector[,paste0(parm,"P")]<=thresholdp

table(collector$originalP<=thresholdp, collector[paste0(parm,"P")]<=thresholdp) %>% fisher.test()

cor.test(collector$originall2FC[idx],collector[,paste0(parm,"l2FC")][idx],
         method = "spearman")

qqplot(y=-log10(collector[,paste0(parm,"P")]), 
       x = -log(runif(nrow(collector))), xlim=c(0,12),ylim=c(0,12),
       col="gray", ylab="", xlab="")
par(new=T)
qqplot(y=-log10(collector$originalP), 
       x = -log(runif(nrow(collector))),xlim=c(0,12),ylim=c(0,12),
       xlab="expected",ylab="observed")
abline(0,1,col="red")
legend("topleft", pch=1, col=c("black", "gray"), legend=c("original", parm))


plot(collector$originall2FC[idx],collector[,paste0(parm,"l2FC")][idx], pch=16, 
     main="log 2 foldchange", ylab=parm, xlab="original")
```

### excluding int_dis

```{r effect_of_w/o_intdis}
### excluding int_dist

modelpar=as.character(design(dds_filt))[2]
modelpar=as.formula(paste("~",gsub("int_dis +", "", modelpar)))

design(workingcopy) = modelpar

workingcopy = DESeq(workingcopy)
parmres=results(workingcopy)

parm="wo.int.dis"
collector[,paste0(parm,"P")] = parmres$pvalue
collector[,paste0(parm,"l2FC")] = parmres$log2FoldChange

idx=collector$originalP<=thresholdp
idx=collector[,paste0(parm,"P")]<=thresholdp

table(collector$originalP<=thresholdp, collector[paste0(parm,"P")]<=thresholdp) %>% fisher.test()

cor.test(collector$originall2FC[idx],collector[,paste0(parm,"l2FC")][idx],
         method = "spearman")

qqplot(y=-log10(collector[,paste0(parm,"P")]), 
       x = -log(runif(nrow(collector))), xlim=c(0,12),ylim=c(0,12),
       col="gray", ylab="", xlab="")
par(new=T)
qqplot(y=-log10(collector$originalP), 
       x = -log(runif(nrow(collector))),xlim=c(0,12),ylim=c(0,12),
       xlab="expected",ylab="observed")
abline(0,1,col="red")
legend("topleft", pch=1, col=c("black", "gray"), legend=c("original", parm))


plot(collector$originall2FC[idx],collector[,paste0(parm,"l2FC")][idx], pch=16,
     main="log 2 foldchange", ylab=parm, xlab="original")

```

## Sensitivity main hit

For the most significant tag of interest (5' of the SLITRK5 gene), we
tested if the group effect is stable if correcting for Ethnicity
(PC1-PC4) or CD associated environmental risk factors.

```{r disect_SLTIRK5}

tophit=which.min(results_Deseq$padj)
methdata=log2_cpm[tophit,]
Probdat=as.data.frame(colData(dds_filt))
Probdat$topHit=methdata[rownames(Probdat)]


model0=as.character(design(dds_filt))[2]
model0=as.formula(gsub("0 +", "topHit ~ 0 + ", model0))

lmres=lm(model0, data=Probdat)
lmrescoeff = as.data.frame(coefficients(summary(lmres)))

totestpar=c("site","PC_1", "PC_2", "PC_3", "PC_4", envFact)

ressens=data.frame(matrix(nrow = length(totestpar)+1, ncol=c(3)))
colnames(ressens) = c("beta", "se", "p.value")
rownames(ressens) = c("original", totestpar)

ressens["original",] = lmrescoeff["groupCD", c("Estimate", "Std. Error", "Pr(>|t|)")]

for( parm in totestpar){
  modelpar=as.character(design(dds_filt))[2]
  modelpar=as.formula(gsub("0", paste0("topHit ~ 0 +",parm), modelpar))
  lmres=lm(modelpar, data=Probdat)
  lmrescoeff = as.data.frame(coefficients(summary(lmres)))
  ressens[parm,] = lmrescoeff["groupCD", c("Estimate", "Std. Error", "Pr(>|t|)")] 
}

modelpar=as.character(design(dds_filt))[2]
modelpar=as.formula(gsub("int_dis +", "", gsub("0", "topHit ~ 0", modelpar)))

lmres=lm(modelpar, data=Probdat)
lmrescoeff = as.data.frame(coefficients(summary(lmres)))
ressens["w/o_int_dis",] = lmrescoeff["groupCD", c("Estimate", "Std. Error", "Pr(>|t|)")] 


a = barplot(height = ressens$beta, 
            ylim=rev(range(c(0,ressens$beta-ressens$se)))*1.3, 
            names.arg = rownames(ressens), col=Set3, border = NA, las=3, 
            ylab="beta[se]", main="Effect sensitvity analysis")

arrows(a,ressens$beta, a, ressens$beta+ressens$se, angle = 90, length = 0.1)
arrows(a,ressens$beta, a, ressens$beta-ressens$se, angle = 90, length = 0.1)

text(a, min(ressens$beta-ressens$se)*1.15, 
     formatC(ressens$p.value), cex=0.6, srt=90)


```

All models are corrected for: <br> `r clinFact`, <br> site is included
as random effect. <br>

original: model defined as `r as.character(model0)[3]` <br>

all other models represent the original model + the variable of interest
<br>

# Real-time PCR validation

## Data loading and parsing

```{r RTPCR}

RefGenes = c("GUSB")
Targets_of_Int = c("SLITRK5", "MIR4500HG")
nreplicates = 3
flagscore=Inf #replication quality error

SamplesMeta=read_xlsx(paste0(Home,"/data/RTrawdata/ZelllinienRNA_femNAT.xlsx"))
as.data.frame(SamplesMeta) -> SamplesMeta
                      
SamplesMeta$Pou=paste("POU", SamplesMeta$Pou)
rownames(SamplesMeta)=SamplesMeta$Pou

SamplesMeta$Group = dds_filt$group[match(SamplesMeta$femNATID, dds_filt$ID_femNAT)]


Files=list.files(paste0(Home,"/data/RTrawdata/"), full.names = T)

Files=Files[grepl("_data",Files)]

Sets=unique(substr(basename(Files), 1,8))

Targets_all=vector()
Samples_all=vector()


geoMean=function(x){
  x=x[!is.na(x)]
  if(length(x)==0)
    return(NA)
  else
    return((prod(x))^(1/length(x)))}

for (Set in Sets){

    Setfiles=Files[grep(Set, Files)]

    for( i in 1:length(Setfiles)){
      tmp=read.table(Setfiles[i], skip=8, header=T, sep="\t", comment.char = "", fill=T)[1:96,]
      tmp=tmp[,c("Sample.Name", "Target.Name","CÑ.")]
      colnames(tmp)=c("Sample.Name", "Target.Name", "CT")
      tmp$Target.Name=gsub("SLITRK5_L", "SLITRK5_", tmp$Target.Name)
      tmp$Target.Name=gsub("VD_", "", tmp$Target.Name)
      tmp$Target.Name=gsub("_", "", tmp$Target.Name)
      tmp$Target.Name=substr(tmp$Target.Name,1, regexpr("#", tmp$Target.Name)-1)
      tmp$CT=as.numeric(tmp$CT)
      
      # set bad replicates to NA 
      tmpmu = tapply(tmp$CT, paste0(tmp$Sample.Name,"_",tmp$Target.Name), mean, na.rm=T)
      tmpsd = tapply(tmp$CT, paste0(tmp$Sample.Name,"_",tmp$Target.Name), sd, na.rm=T)
      for (corr in which(tmpsd>flagscore)){
        index=unlist(strsplit(names(tmpmu)[corr], "_"))
        tmp[which(tmp$Sample.Name==index[1] & tmp$Target.Name==index[2]),"CT"] = NA
      }
      assign(paste0("tmp_",Set,"_",i),tmp)
    }
    
    tmp=do.call("rbind", mget(apropos(paste0("tmp_",Set))))
    tmp=tmp[which(!(tmp$Sample.Name==""|is.na(tmp$Sample.Name))), ]
    tmp=tmp[!tmp$Sample.Name=="NTC",]
    
    Samples=unique(tmp$Sample.Name)
    Targets=unique(tmp$Target.Name)
    Samples_all=unique(c(Samples_all, Samples))
    Targets_all=unique(c(Targets_all, Targets))
    
    Reform=data.frame(matrix(NA, nrow=length(Samples), ncol=length(Targets)*nreplicates))
    
    colnames(Reform)=paste0(rep(Targets, each=3), letters[1:nreplicates])
    rownames(Reform)=Samples
    
    for (i in Samples) {
      #print(i)
      for (j in Targets){
        Reform[i,grep(j, colnames(Reform))]=tmp[tmp$Sample.Name==i & tmp$Target.Name==j,"CT"]
      }
    }
    
    HK=colnames(Reform)[grep(paste0(RefGenes, collapse="|"),colnames(Reform))]
    
    GMHK=apply(Reform[,HK], 1, geoMean)
    tmp2=Reform-GMHK
    
    assign(paste0(Set,"_dCT"), tmp2)
    rm(list=c(apropos("tmp"), "Reform", "GMHK"))
    
}

Samples_all=unique(Samples_all)
Targets_all = unique(Targets_all)

mergedCTtable=data.frame(matrix(NA,ncol=length(Targets_all)*nreplicates, nrow=length(Samples_all)))
colnames(mergedCTtable)=paste0(rep(unique(Targets_all), each=nreplicates), letters[1:nreplicates])
rownames(mergedCTtable)=Samples_all

CTobj=apropos("_dCT")

for( obj in CTobj){
  DF=get(obj)
  for(k in colnames(DF)){
    for(l in rownames(DF)){
      mergedCTtable[l,k]=DF[l,k]
    }
  }
}

CTmeans=colMeans(mergedCTtable, na.rm = T)

meanvec=tapply(CTmeans,gsub(paste0(letters[1:nreplicates],collapse="|"),"",names(CTmeans)), mean, na.rm=T)
meanvec = rep(meanvec, each=nreplicates)
names(meanvec) = paste0(names(meanvec), letters[1:nreplicates])
meanvec=meanvec[colnames(mergedCTtable)]
ddCT=apply(mergedCTtable,1, function(x){x-meanvec}) 

FC=2^-ddCT


SamplesMeta$inset=F
SamplesMeta$inset[SamplesMeta$Pou %in% colnames(FC)]=T

SamplesMeta=SamplesMeta[SamplesMeta$inset,]

CTRLCASEsorter=c(which(SamplesMeta$Group=="CTRL"),which(SamplesMeta$Group=="CD"))
SamplesMeta = SamplesMeta[CTRLCASEsorter, ]

searcher=paste0(Targets_of_Int, collapse = "|")

FC = FC[grepl(searcher, rownames(FC)),SamplesMeta$Pou]

MuFC=apply(FC, 2, function(x){tapply(log2(x), gsub("a|b|c","",rownames(FC)), mean, na.rm=T)})
SDFC=apply(FC, 2, function(x){tapply(log2(x), gsub("a|b|c","",rownames(FC)), sd, na.rm=T)})


```

## plot relative expression by samples

```{r RTPCR_step3plot}
pdf(paste0(Home, "/output/barplots.pdf"))

for(i in Targets_of_Int){
  if(any(!is.na(MuFC[i,]))){
  a=barplot(unlist(MuFC[i,]), col=as.numeric(SamplesMeta[colnames(MuFC),"Group"])+1, main=i, las=3, 
            ylim=c(-max(abs(MuFC[i,])*1.2, na.rm=T), (max(abs(MuFC[i,])*1.2, na.rm=T))))
  arrows(a, MuFC[i,], a, MuFC[i,]+SDFC[i,], angle = 90, length = 0.1)
  arrows(a, MuFC[i,], a, MuFC[i,]-SDFC[i,], angle = 90, length = 0.1)
  legend("topleft", c("case", "control"), col=c(1,2), pch=15, bty="n")
  } else {
    plot(0,0, type="n", main=paste(i, "not detected"))
  }
}

dev.off()

for(i in Targets_of_Int){
  if(any(!is.na(MuFC[i,]))){
  a=barplot(unlist(MuFC[i,]), col=as.numeric(SamplesMeta[colnames(MuFC),"Group"])+1, main=i, las=3, 
            ylim=c(-max(abs(MuFC[i,])*1.2, na.rm=T), (max(abs(MuFC[i,])*1.2, na.rm=T))))
  arrows(a, MuFC[i,], a, MuFC[i,]+SDFC[i,], angle = 90, length = 0.1)
  arrows(a, MuFC[i,], a, MuFC[i,]-SDFC[i,], angle = 90, length = 0.1)
  legend("topleft", c("case", "control"), col=c(1,2), pch=15, bty="n")
  } else {
    plot(0,0, type="n", main=paste(i, "not detected"))
  }
}


```

## compare across groups

```{r RTPCR_step4 test}
sink(paste0(Home, "/output/ResultsgroupComp.txt"))

Group=SamplesMeta$Group
for(i in Targets_of_Int){
  print(i)
  print(summary(try(lm(unlist(MuFC[i,])~Group))))
  print(t.test(unlist(MuFC[i,])~Group))
}

sink()


SamplesMeta$femNATID2=paste0("ID_",gsub("-","_",SamplesMeta$femNATID))

SamplesMeta=SamplesMeta[SamplesMeta$Pou %in% colnames(MuFC),]
MuFC=MuFC[,SamplesMeta$Pou]

TPM4RNA=selEpitpm[,SamplesMeta$femNATID2]
colnames(TPM4RNA)=SamplesMeta$Pou

tags=list()


Targets=Targets_of_Int


sigtags=which(restab$padj<=0.05)

tagsOI=grep(paste0(Targets, collapse = "|"),selEpiMeta$gene)

sigtagsOI = tagsOI[tagsOI %in% sigtags]

fintagsOI=data.frame(tags=sigtagsOI, gene=selEpiMeta[sigtagsOI,"gene"])

#Targ=Targets[1]
#tag=tags[1]
```

## compare against methylation level

```{r RTPCR_step4 test_plot, fig.width=15,fig.height=8}

pdf(paste0(Home,"/output/RNAvsMETplots.pdf"), width = 15, height = 8)

MuFC=as.data.frame(MuFC)
MuFCsel=MuFC[Targets,]
par(mar=c(5,5,5,3), mfrow=c(1,2))
for (Targ in Targets){
  tags=fintagsOI$tags[grep(Targ, fintagsOI$gene)]
  for (tag in tags){
    data=data.frame(tpm=unlist(TPM4RNA[tag,SamplesMeta$Pou]) , 
                    RT=unlist(MuFCsel[grep(Targ, rownames(MuFCsel)),SamplesMeta$Pou]))
    plot(data$tpm,data$RT, 
         xlab="methylation tpm", 
         ylab = "mRNA log2FC vs mean", 
         ylim=c(-3,3),
         col=4-as.numeric(SamplesMeta$Group), 
         pch=as.numeric(SamplesMeta$Group)+14,
         main=paste(tag, "Meth vs mRNA Expr", Targ))
    legend("topleft", c("control", "case"), pch=c(15,16), col=c(3,2), bty="n")
    a=lm(RT~tpm, data)
    b=summary(a)
    abline(a, col="blue")

    SperCor=cor(data$RT,data$tpm,use = "c", method = "spearman")
    mtext(3, text = paste("beta = ", round(coefficients(a)[2],2), 
                          "; se =", round(b$coefficients[2,2],2), 
                          "; pvalue = ", round(b$coefficients[2,4],3), 
                          "; sperman cor = ", round(SperCor,3)))
  }
}

dev.off()

MuFC=as.data.frame(MuFC)
MuFCsel=MuFC[Targets,]
par(mar=c(5,5,5,3), mfrow=c(1,2))
for (Targ in Targets){
  tags=fintagsOI$tags[grep(Targ, fintagsOI$gene)]
  for (tag in tags){
    data=data.frame(tpm=unlist(TPM4RNA[tag,SamplesMeta$Pou]) , 
                    RT=unlist(MuFCsel[grep(Targ, rownames(MuFCsel)),SamplesMeta$Pou]))
    plot(data$tpm,data$RT, 
         xlab="methylation tpm", 
         ylab = "mRNA log2FC vs mean", 
         ylim=c(-3,3),
         col=4-as.numeric(SamplesMeta$Group), 
         pch=as.numeric(SamplesMeta$Group)+14,
         main=paste(tag, "Meth vs mRNA Expr", Targ))
    legend("topleft", c("control", "case"), pch=c(15,16), col=c(3,2), bty="n")
    a=lm(RT~tpm, data)
    b=summary(a)
    abline(a, col="blue")
    SperCor=cor(data$RT,data$tpm,use = "c", method = "spearman")
    mtext(3, text = paste("beta = ", round(coefficients(a)[2],2), 
                          "; se =", round(b$coefficients[2,2],2), 
                          "; pvalue = ", round(b$coefficients[2,4],3), 
                          "; Spearman cor = ", round(SperCor,3)))
  }
}

```

# Systems Analysis

## Genomic feature enrichment

Significant loci with a p-value \<= `r thresholdp` and a absolute log2
fold-change lager `r thresholdLFC` were tested for enrichment in
annotated genomic feature using fisher exact test.

```{r genomiclocation, fig.width=5, fig.height=5}

Ranges=rowData(dds_filt)

TotTagsofInterest=sum(Ranges$WaldPvalue_groupCD<=thresholdp & abs(Ranges$groupCD)>thresholdLFC)

Resall=data.frame()
index = Ranges$WaldPvalue_groupCD<=thresholdp& abs(Ranges$groupCD)>thresholdLFC
for (feat in unique(Ranges$feature)){
  tmp=table(Ranges$feature == feat, signif=index)
  resfish=fisher.test(tmp)
  res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
  Resall = rbind(Resall, res)
}
tmp=table(Ranges$tf_binding!="", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resall = rbind(Resall, res)
tmp=table(Ranges$cpg=="cpg", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resall = rbind(Resall, res)
colnames(Resall)=c("OR", "CI95L", "CI95U", "P")
rownames(Resall)=c(unique(Ranges$feature), "TF-binding", "CpG-island")
Resall$Beta = log(Resall$OR)
Resall$SE = (log(Resall$OR)-log(Resall$CI95L))/1.96
Resall$Padj=p.adjust(Resall$P, method = "bonferroni")

Resdown=data.frame()
index = Ranges$WaldPvalue_groupCD<=thresholdp & Ranges$groupCD<thresholdLFC
for (feat in unique(Ranges$feature)){
  tmp=table(Ranges$feature == feat, signif=index)
  resfish=fisher.test(tmp)
  res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
  Resdown = rbind(Resdown, res)
}
tmp=table(Ranges$tf_binding!="", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resdown = rbind(Resdown, res)
tmp=table(Ranges$cpg=="cpg", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resdown = rbind(Resdown, res)
colnames(Resdown)=c("OR", "CI95L", "CI95U", "P")
rownames(Resdown)=c(unique(Ranges$feature), "TF-binding", "CpG-island")
Resdown$Beta = log(Resdown$OR)
Resdown$SE = (log(Resdown$OR)-log(Resdown$CI95L))/1.96
Resdown$Padj=p.adjust(Resdown$P, method = "bonferroni")

Resup=data.frame()
index = Ranges$WaldPvalue_groupCD<=thresholdp & Ranges$groupCD>thresholdLFC
for (feat in unique(Ranges$feature)){
  tmp=table(Ranges$feature == feat, signif=index)
  resfish=fisher.test(tmp)
  res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
  Resup = rbind(Resup, res)
}
tmp=table(Ranges$tf_binding!="", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resup = rbind(Resup, res)
tmp=table(Ranges$cpg=="cpg", signif=index)
resfish=fisher.test(tmp)
res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
Resup = rbind(Resup, res)
colnames(Resup)=c("OR", "CI95L", "CI95U", "P")
rownames(Resup)=c(unique(Ranges$feature), "TF-binding", "CpG-island")
Resup$Beta = log(Resup$OR)
Resup$SE = (log(Resup$OR)-log(Resup$CI95L))/1.96
Resup$Padj=p.adjust(Resup$P, method = "bonferroni")

multiORplot(Resall, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="All diff. methylated loci")
multiORplot(Resup, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="hypomethylated loci")
multiORplot(Resdown, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="Hypermethylated loci")

pdf(paste0(Home, "/output/functional_Enrichemnt.pdf"))
multiORplot(Resall, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="All diff. methylated loci")
multiORplot(Resup, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="hypomethylated loci")
multiORplot(Resdown, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", pheno="Hypermethylated loci")
dev.off()


```

## GO-term Enrichment

Significant loci and differentially methylated regions with a p-value
\<= `r thresholdp` and an absolute log2 fold-change lager
`r thresholdLFC` were tested for enrichment among GO-terms Molecular
Function, Cellular Compartment and Biological Processes, KEGG pathways,
Transcription factor Binding sites, Human Protein Atlas Tissue
Expression, Human Phenotypes.

```{r go_term, fig.height=15, fig.width=9}

getGOresults = function(geneset, genereference){
  resgo = gost(geneset, organism = "hsapiens",
               correction_method = "g_SCS",
               domain_scope = "custom",
               sources = c("GO:BP", "GO:MF", "GO:CC"),
               custom_bg = genereference)
  if(length(resgo) != 0){
    return(resgo)
  } else {
    print("no significant results")
    return(NULL)
  }  
}

gene_univers = getuniquegenes(as.data.frame(rowRanges(dds_filt))$gene)


idx = (results_Deseq$pvalue <= thresholdp & 
         (abs(results_Deseq$log2FoldChange) > thresholdLFC))

genes_reg = getuniquegenes(as.data.frame(rowRanges(dds_filt))$gene[idx])


dmr_genes = unique(resultsdmr_table$name[resultsdmr_table$p.value<=thresholdp & 
                   abs(resultsdmr_table$value)>=thresholdLFC])


Genes_of_interset = list("01_dmregions" = dmr_genes,  
                         "02_dmtag" = genes_reg
                         )

gostres = getGOresults(Genes_of_interset, gene_univers)

gostplot(gostres, capped = TRUE, interactive = T)
p = gostplot(gostres, capped = TRUE, interactive = F)

toptab = gostres$result

pp = publish_gostplot(p, filename = paste0(Home,"/output/gostres.pdf"))
write.xlsx2(toptab, file = paste0(Home,"/output/GOres.xlsx"), sheetName = "GO_enrichment")


```

# Brain Developmental Processes Enrichment tests

Gene sets identified to be deferentially methylated with a p-value \<=
`r thresholdp` and an absolute log2 fold-change larger `r thresholdLFC`
were tested for enrichment among gene-modules coregulated during Brain
expression.

## Kang Modules

```{r brainModules, fig.height=12, fig.width=5}

# define Reference Universe 

KangUnivers<- read.table(paste0(Home,"/data/KangUnivers.txt"), sep="\t", header=T)
colnames(KangUnivers)<-c("EntrezId","Symbol")

Kang_genes<-read.table(paste0(Home,"/data/Kang_dataset_genesMod_version2.txt"),sep="\t",header=TRUE)

#3)Generate Gene universe to be used for single gene lists
tmp=merge(KangUnivers,Kang_genes,by.y="EntrezGene",by.x="EntrezId",all=TRUE) #18826
KangUni_Final<-tmp[duplicated(tmp$EntrezId)==FALSE,] #18675


# Local analysis gene universe
Annotation_list<-data.frame(Symbol = gene_univers)

# match modules 
Annotation_list$Module = Kang_genes$Module[match(Annotation_list$Symbol,Kang_genes$symbol)]

# check if overlapping in gene universes
Annotation_list$univers = Annotation_list$Symbol %in% KangUni_Final$Symbol

# drop duplicates 
Annotation_list = Annotation_list[duplicated(Annotation_list$Symbol)==FALSE,]

# selct only genes that have been detected on both datasets
Annotation_list = Annotation_list[Annotation_list$univers==T,] 

# final reference 
UniversalGeneset=Annotation_list$Symbol

# define Gene lists to test 
# sort and order Modules to be tested

Modules=unique(Annotation_list$Module)
Modules = Modules[! Modules %in% c(NA, "")]
Modules = Modules[order(as.numeric(gsub("M","",Modules)))]

GL_all=list()

for(i in Modules){
  GL_all[[i]]=Annotation_list$Symbol[Annotation_list$Module%in%i]
}
GL_all[["M_all"]]=Kang_genes$symbol[Kang_genes$Module %in% Modules]


GOI1 = Genes_of_interset

Resultsall=list()
for(j in names(GOI1)){
  Res = data.frame()
  for(i in names(GL_all)){
    Modulegene=GL_all[[i]]
    Factorgene=GOI1[[j]]
    Testframe<-fisher.test(table(factor(UniversalGeneset %in% Factorgene,levels=c("TRUE","FALSE")),
                                 factor(UniversalGeneset %in% Modulegene,levels=c("TRUE","FALSE"))))
    beta=log(Testframe$estimate)
    Res[i, "beta"] =beta
    Res[i, "SE"]=abs(beta-log(Testframe$conf.int[1]))/1.96
    Res[i, "Pval"]=Testframe$p.value
    Res[i, "OR"]=(Testframe$estimate)
    Res[i, "ORL"]=(Testframe$conf.int[1])
    Res[i, "ORU"]=(Testframe$conf.int[2])
  }
  Res$Padj = p.adjust(Res$Pval, method = "bonferroni")
  Resultsall[[j]] = Res
  
}
par(mfrow = c(2,1))
for (i in names(Resultsall)){ 
  multiORplot(datatoplot = Resultsall[[i]], pheno=i)
}

par(mfrow = c(1,1))
pdf(paste0(Home, "/output/BrainMod_Enrichemnt.pdf"))
for (i in names(Resultsall)){
  multiORplot(datatoplot = Resultsall[[i]], pheno=i)
}
dev.off()


Modsig = c()

for(r in names(Resultsall)){
  a=rownames(Resultsall[[r]])[Resultsall[[r]]$Padj<=0.05]
  Modsig = c(Modsig,a)
}




```

## Brain espresseion heatmaps

```{r modplots}

# show brains and expression
Modsig2=unique(Modsig[Modsig!="M_all"])

load(paste0(Home,"/data/Kang_DataPreprocessing.RData")) #Load the Kang expression data of all genes 
datExprPlot=matriz #Expression data of Kang loaded as Rdata object DataPreprocessing.RData


Genes = GL_all[names(GL_all)!="M_all"]

Genes_expression<-list()

pcatest<-list()
for (i in names(Genes)){
  Genes_expression[[i]]<-matriz[,which(colnames(matriz) %in% Genes[[i]])]
  pcatest[[i]]=prcomp(t(as.matrix(Genes_expression[[i]])),retx=TRUE)
}

# PCA test
PCA<-data.frame(pcatest[[1]]$rotation)
PCA$donor_name<-rownames(PCA)
PC1<-data.frame(PCA[,c(1,ncol(PCA))])

#Combining the age with expression data
list <- strsplit(sampleInfo$age, " ")
library("plyr")
df <- ldply(list)
colnames(df) <- c("Age", "time")

sampleInfo<-cbind(sampleInfo[,1:9],df)
sampleInfo$Age<-as.numeric(sampleInfo$Age)

sampleInfo$period<-ifelse(sampleInfo$time=="pcw",sampleInfo$Age*7,ifelse(sampleInfo$time=="yrs",sampleInfo$Age*365+270,ifelse(sampleInfo$time=="mos",sampleInfo$Age*30+270,NA)))

#We need it just for the donor names

PCA_matrix<-merge.with.order(PC1,sampleInfo,by.y="SampleID",by.x="donor_name",keep_order=1)

#Select which have phenotype info present 
matriz2<-matriz[which(rownames(matriz) %in% PCA_matrix$donor_name),]
FactorGenes_expression<-list()
#Factors here mean modules
for (i in names(Genes)){
  FactorGenes_expression[[i]]<-matriz2[,which(colnames(matriz2) %in% Genes[[i]])]
}


FactorseGE<-list()
for (i in names(Genes)){
  FactorseGE[[i]]<-FactorGenes_expression[[i]]
}

allModgenes=NULL
colors=vector()
for ( i in names(Genes)){
  allModgenes=cbind(allModgenes,FactorseGE[[i]])
  colors=c(colors, rep(i, ncol(FactorseGE[[i]])))
}


lengths=unlist(lapply(FactorGenes_expression, ncol), use.names = F)

MEorig=moduleEigengenes(allModgenes, colors)

PCA_matrixfreeze=PCA_matrix

index=!PCA_matrix$structure_acronym %in% c("URL", "DTH", "CGE","LGE", "MGE",  "Ocx", "PCx", "M1C-S1C","DIE", "TCx", "CB")
PCA_matrix=PCA_matrix[index,]
ME = MEorig$eigengenes[index,]
matsel = matriz2[index,]

colnames(ME) = gsub("ME", "", colnames(ME))

timepoints=seq(56,15000, length.out=1000)
matrix(c("CB", "THA", "CBC", "MD"), ncol=2 ) -> cnm


brainheatmap=function(Module){
  MEmod=ME[,Module]
  toplot=data.frame(matrix(NA, nrow=length(table(PCA_matrix$structure_acronym)), ncol=998))
  rownames(toplot)=unique(PCA_matrix$structure_acronym)
  target <- c("OFC", "DFC", "VFC", "MFC","M1C","S1C","IPC","A1C","STC","ITC","V1C","HIP","AMY","STR","MD","CBC")
  toplot<-toplot[c(6,2,8,5,11,12,10,9,7,4,14,3,1,13,16,15),]
  
  
  for ( i in unique(PCA_matrix$structure_acronym)){
    index=PCA_matrix$structure_acronym==i
    LOESS=loess(MEmod[index]~PCA_matrix$period[index])
    toplot[i,]=predict(LOESS,newdata = round(exp(seq(log(56),log(15000), length.out=998)),2))
    colnames(toplot)[c(1,77,282,392,640,803,996)]<-
      c("1pcw","21pcw","Birth","1.3years","5.4years","13.6years","40.7years")
  }
  
  
  
  cols=viridis(100)
  labvec <- c(rep(NA, 1000))
  
  labvec[c(1,77,282,392,640,803,996)] <- c("1pcw","21pcw","Birth","1.3years","5.4years","13.6years","40.7years")
  
  
  toplot<-toplot[,1:998]
  date<-c(1:998)
  dateY<-paste0(round(date/365,2),"_Years")
  
  names(toplot)<-dateY
  
  par(xpd=FALSE) 
  heatmap.2(as.matrix(toplot), col = cols, 
            main=Module,
            trace = "none", 
            na.color = "grey",
            Colv = F, Rowv = F,
            labCol = labvec,
            #breaks = seq(-0.1,0.1, length.out=101),
            symkey = T,
            scale = "row",
            key.title = "",
            dendrogram = "none",
            key.xlab = "eigengene",
            density.info = "none",
            #main=paste("Module",1),
            srtCol=90,
            tracecol = "none", 
            cexRow = 1,
            add.expr=eval.parent(abline(v=282),
                                 axis(1,at=c(1,77,282,392,640,803,996),
                                      labels =FALSE)),cexCol = 1)
  
 
}

brainheatmap_gene=function(Genename){
  MEmod=matsel[,Genename]
  toplot=data.frame(matrix(NA, nrow=length(table(PCA_matrix$structure_acronym)), ncol=998))
  rownames(toplot)=unique(PCA_matrix$structure_acronym)
  target <- c("OFC", "DFC", "VFC", "MFC","M1C","S1C","IPC","A1C","STC","ITC","V1C","HIP","AMY","STR","MD","CBC")
  toplot<-toplot[c(6,2,8,5,11,12,10,9,7,4,14,3,1,13,16,15),]


  for ( i in unique(PCA_matrix$structure_acronym)){
    index=PCA_matrix$structure_acronym==i
    LOESS=loess(MEmod[index]~PCA_matrix$period[index])
    toplot[i,]=predict(LOESS,newdata = round(exp(seq(log(56),log(15000), length.out=998)),2))
    colnames(toplot)[c(1,77,282,392,640,803,996)]<-
      c("1pcw","21pcw","Birth","1.3years","5.4years","13.6years","40.7years")
  }



  cols=viridis(100)
  labvec <- c(rep(NA, 1000))

  labvec[c(1,77,282,392,640,803,996)] <- c("1pcw","21pcw","Birth","1.3years","5.4years","13.6years","40.7years")


  toplot<-toplot[,1:998]
  date<-c(1:998)
  dateY<-paste0(round(date/365,2),"_Years")

  names(toplot)<-dateY

  par(xpd=FALSE)
  heatmap.2(as.matrix(toplot), col = cols,
            main=Genename,
            trace = "none",
            na.color = "grey",
            Colv = F, Rowv = F,
            labCol = labvec,
            #breaks = seq(-0.1,0.1, length.out=101),
            symkey = F,
            scale = "none",
            key.title = "",
            dendrogram = "none",
            key.xlab = "eigengene",
            density.info = "none",
            #main=paste("Module",1),
            #srtCol=90,
            tracecol = "none",
            cexRow = 1,
            add.expr=eval.parent(abline(v=282),
                                 axis(1,at=c(1,77,282,392,640,803,996),
                                      labels =FALSE))
            ,cexCol = 1)
}


brainheatmap_gene("SLITRK5")
for(Module in Modsig2){
  brainheatmap(Module)
}

pdf(paste0(Home, "/output/Brain_Module_Heatmap.pdf"))

brainheatmap_gene("SLITRK5")
for(Module in Modsig2){
  brainheatmap(Module)
}
dev.off()




```

# Risk Factor Mediation Analysis

## Risk factor loading and correlation plots

```{r SEM, , fig.width=15, fig.height=8}
dropfact=c("site", "0", "group")

modelFact=strsplit(as.character(design(dds_filt))[2], " \\+ ")[[1]]


Patdata=as.data.frame(colData(dds_filt))

load(paste0(Home, "/output/envFact.RData"))

envFact=envFact[!envFact %in% dropfact] 
modelFact=modelFact[!modelFact %in% dropfact] 

EpiMarker = c()

# TopHit
Patdata$Epi_TopHit=log2_cpm[base::which.min(results_Deseq$pvalue),]

# 1PC of all diff met
tmp=glmpca(log2_cpm[base::which(results_Deseq$pvalue<=thresholdp),], 1)

Patdata$Epi_all= tmp$factors$dim1
  
EpiMarker = c(EpiMarker, "Epi_TopHit", "Epi_all")

#Brain Modules

Epitestset=GL_all[Modsig]

for(n in names(Epitestset)){
  index=gettaglistforgenelist(genelist = Epitestset[[n]], dds_filt)
  index = base::intersect(index, base::which(results_Deseq$pvalue<=thresholdp))
  # get eigenvalue
  epiname=paste0("Epi_",n)
  tmp=glmpca(log2_cpm[index,], 1)
  Patdata[,epiname]= tmp$factors$dim1
  EpiMarker = c(EpiMarker, epiname)
}

cormat = cor(apply(Patdata[,c("group", envFact, modelFact, EpiMarker)] %>% mutate_all(as.numeric), 2, minmax_scaling),
             use = "pairwise.complete.obs")

par(mfrow=c(1,2))
corrplot(cormat, main="correlations")
corrplot(cormat, order = "hclust", main="correlations ordered")
```

## SEM analysis

```{r SEM_modelling, , fig.width=15, fig.height=8}
fullmodEnv=paste(unique(envFact,modelFact), sep = "+", collapse = "+")

Dataset = Patdata[,c("group", envFact, modelFact,EpiMarker)]

model = "
Epi~0+a*Matsmk+b*Matagg+c*FamScore+d*EduPar+e*n_trauma+Age+int_dis+medication+contraceptives+cigday_1+V8

group~f*Matsmk+g*Matagg+h*FamScore+i*EduPar+j*n_trauma+Age+int_dis+medication+contraceptives+cigday_1+V8+z*Epi

#direct
directMatsmk := f
directMatagg := g
directFamScore := h
directEduPar := i
directn_trauma := j

#indirect
EpiMatsmk := a*z
EpiMatagg := b*z
EpiFamScore := c*z
EpiEduPar := d*z
Epin_trauma := e*z

total := f + g + h + i + j + (a*z)+(b*z)+(c*z)+(d*z)+(e*z)

"

Netlist = list()


nothing = function(x){return(x)}

for (marker in EpiMarker) {
  
  Dataset$Epi = Dataset[,marker]
  Datasetscaled = Dataset %>% mutate_if(is.numeric, minmax_scaling)
  Datasetscaled = Datasetscaled %>% mutate_if(is.factor,ordered)

  fit<-lavaan(model,data=Datasetscaled, estimator="DWLS")
  
  sink(paste0(Home,"/output/SEM_summary_group",marker,".txt"))
  summary(fit)
  print(fitMeasures(fit))
  print(parameterEstimates(fit))
  sink()
  
  cat("############################\n")
  cat("############################\n")
  cat(marker, "\n")
  cat("############################\n")
  cat("############################\n")
  
  cat("##Mediation Model ##\n")
  summary(fit)
  cat("\n")
  #print(fitMeasures(fit))
  cat("\n")
  #print(parameterEstimates(fit))
  cat("\n")
  
  #SOURCE FOR PLOT https://stackoverflow.com/questions/51270032/how-can-i-display-only-significant-path-lines-on-a-path-diagram-r-lavaan-sem
  restab=lavaan::parameterEstimates(fit) %>% dplyr::filter(!is.na(pvalue)) %>% 
    arrange(desc(pvalue)) %>% mutate_if("is.numeric","round",3) %>% 
    dplyr::select(-ci.lower,-ci.upper,-z)
  
  pvalue_cutoff <- 0.05
  
  obj <- semPlot:::semPlotModel(fit)
  original_Pars <- obj@Pars
  
  
  print(original_Pars)
  
  
  check_Pars <- obj@Pars %>% dplyr:::filter(!(edge %in% c("int","<->") | lhs == rhs)) # this is the list of parameter to sift thru
  keep_Pars <- obj@Pars %>% dplyr:::filter(edge %in% c("int","<->") | lhs == rhs) # this is the list of parameter to keep asis
  
  test_against <- lavaan::parameterEstimates(fit) %>% dplyr::filter(pvalue < pvalue_cutoff, rhs != lhs)
  
  # for some reason, the rhs and lhs are reversed in the standardizedSolution() output, for some of the values
  # I'll have to reverse it myself, and test against both orders
  
  test_against_rev <- test_against %>% dplyr::rename(rhs2 = lhs, lhs = rhs) %>% dplyr::rename(rhs = rhs2)
  
  checked_Pars <-
    check_Pars %>% semi_join(test_against, by = c("lhs", "rhs")) %>% bind_rows(
      check_Pars %>% semi_join(test_against_rev, by = c("lhs", "rhs"))
    )
  
  obj@Pars <- keep_Pars %>% bind_rows(checked_Pars) %>% 
    mutate_if("is.numeric","round",3) %>% 
    mutate_at(c("lhs","rhs"),~gsub("Epi", marker,.))
  
  obj@Vars = obj@Vars %>% mutate_at(c("name"),~gsub("Epi", marker,.))
  
  DF = obj@Pars
  DF = DF[DF$lhs!=DF$rhs,]
  DF = DF[abs(DF$est)>0.1,]
  
  DF = DF[DF$edge == "~>",] # only include directly modelled effects in figure 
  
  nodes <- data.frame(id=obj@Vars$name, label = obj@Vars$name)
  nodes$color<-Dark8[8]
  nodes$color[nodes$label == "group"] = Dark8[3]
  nodes$color[nodes$label == marker] = Dark8[4]
  nodes$color[nodes$label %in% envFact] = Dark8[5]
  
  if(nrow(DF)>0){
  edges <- data.frame(from = DF$lhs, 
                      to = DF$rhs, 
                      width=abs(DF$est), 
                      arrows ="to")
  
  edges$dashes = F
  edges$label =  DF$est
  edges$color=c("firebrick", "forestgreen")[1:2][factor(sign(DF$est), levels=c(-1,0,1),labels=c(1,2,2))]
  edges$width=2
  }
  else {edges = data.frame(from=NULL, to = NULL)}
  
  cexlab = 18
  plotnet<- visNetwork(nodes, edges,  
                       main=list(text=marker,
                                 style="font-family:arial;font-size:20px;text-align:center"),
                       submain=list(text="significant paths", 
                                    style="font-family:arial;text-align:center")) %>% 
    visEdges(arrows =list(to = list(enabled = TRUE, scaleFactor = 0.7)),
                       font=list(size=cexlab, style="font-family:arial;text-align:center")) %>%
    visNodes(font=list(size=cexlab, style="font-family:arial;text-align:center")) %>%
    
    visPhysics(enabled = T, solver = "forceAtlas2Based")
  
  
  Netlist[[marker]] = plotnet
  htmlfile = paste0(Home,"/output/SEMplot_",marker,".html")
  visSave(plotnet, htmlfile)
  webshot(paste0(Home,"/output/SEMplot_",marker,".html"), zoom = 1, 
          file = paste0(Home,"/output/SEMplot_",marker,".png"))
  
}

```

```{r plotsems_prep, results="asis"}

rmd_paths <-paste0(tempfile(c(names(Netlist))),".Rmd")
names(rmd_paths) <- names(Netlist)

for (n in names(rmd_paths)) {
    sink(file = rmd_paths[n])
    cat("  \n",
        "```{r, echo = FALSE}",
            "Netlist[[n]]",
        "```",
        sep = "  \n")
    sink()
}

```

## Interactive SEM plots

Only direct effects with a significant standardized effect of p\<0.05
are shown.

```{r plotsems_show_interactive, results='asis'}
    for (n in names(rmd_paths)) {
        cat(knitr::knit_child(rmd_paths[[n]],
                              quiet= TRUE))
        file.remove(rmd_paths[[n]])
    }
```
